Add ability to specify the delimiter used for displaying selection as a string.
This change introduces a new configuration parameter "resultAsStringDelimiter" to allow the user to specify, if they so choose, a text delimiter other than the default comma. I needed this change as I was working with linguistic data, and preferred to see the selections as strings, but the commas were causing confusion. This allows me to pass an empty string '' so no delimiter is used, but any text string can be set. It also falls back gracefully in that if resultAsString is true, but no special delimiter is specified at all, the original default of a comma is used.