Fixes for older glitches (ms-trigger, ms-sel-ctn)
The first out of these two rules reduces the unnecessary left margin of the selection container, causing the selection block to look evenly padded from the top, left and bottom edges of the component.
The second rule fixes a glitch where the trigger covers the top-right and bottom-right rounded corners of the component, while on :hover. It basically shortens the trigger and adds rounded borders to it on hover.