Stop errors
Not sure if this is the best way to do this, but I have a situation where we use magic suggest to capture a click event within the magic suggest ui, then do an action from our app. This global listener on body is then throwing errors on the first click after the click in magic suggest. There is no guarantee that `e.target` is a string as far as I can tell. It would seem to me a `$('body').on(someNamespacedEventThatCouldBeCleanedUp, 'click')` would be a better approach possibly?